DeepMol/requirements.txt
rdkit==2023.9.6
smilespe==0.0.3
seaborn
torch==2.5.1
dgl<=2.0.0
joblib
pillow
h5py
shap==0.44.0
mol2vec @ git+https://github.com/samoturk/mol2vec#egg=mol2vec
boruta==0.4.3
tensorflow==2.15.0
chembl_structure_pipeline==1.1.0
scikit-learn<1.6
imblearn
umap-learn==0.5.1
kneed==0.8.2
kaleido==0.2.1
plotly
cached_property==1.5.2
numpy
dgllife==0.3.2
dill==0.3.6
optuna==4.1.0
tensorflow-probability==0.25.0
torch_geometric==2.6.1
scikit-multilearn==0.2.0
scikeras==0.12.0
pydantic==2.6.4
timeout_decorator==0.5.0
IPython
deepchem==2.8.0
pandas<=2.0.0
deepmol